+++ backend/tasks_lib/entities/group_vdbs_task_msg.py
@@ -1,0 +1,5 @@
+from dataclasses import dataclass
+
+@dataclass
+class GroupVDBsTasksMsg:
+    gvt_id: int